1) Deleted @Data from entities 2) Changed Long to double 3) login do not changed 4) Storage -> Atm, CashMachine -> Changed logic, Storage deleted 5) Bill is a storage now 6) Corrected 7) Method deleted 8) Changed 9) Method is ok 10) Changed names of 2 methods 11) Changed to "/add" 12) Added implementation

+++ src/main/java/bank/atm/dto/request/AccountRequestDto.java
@@ -3,9 +3,9 @@ package bank.atm.dto.request;
 public class AccountRequestDto {
     private Long accountNumber;
     private Long userId;
-    private Long money;
+    private double money;
 
-    public Long getMoney() {
+    public double getMoney() {
         return money;
     }
